NewsPortal/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
dill==0.3.8
Django==5.0.7
django-allauth==64.0.0
django-apscheduler==0.6.2
django-filter==24.3
django-modeltranslation==0.19.5
djangorestframework==3.15.2
idna==3.7
isort==5.13.2
jwt==1.3.1
kombu==5.3.7
Markdown==3.6
mccabe==0.7.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
pylint==3.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
